% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOSversion.R
\name{GEOSversion}
\alias{GEOSversion}
\title{Version of GEOS library}
\usage{
GEOSversion()
}
\value{
character string
}
\description{
Returns a string of the GEOS library version.
}
\details{
On Windows this will be the rwinlib version on CRAN, not a system GEOS.
}
\examples{
GEOSversion()
}
