% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libgeos-package.R
\docType{package}
\name{libgeos-package}
\alias{libgeos}
\alias{libgeos-package}
\title{libgeos: Open Source Geometry Engine C API}
\description{
Provides access to the GEOS C API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/libgeos}
  \item Report bugs at \url{https://github.com/paleolimbot/libgeos/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
