% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libgeos-package.R
\name{libgeos_version}
\alias{libgeos_version}
\title{GEOS version information}
\usage{
libgeos_version()
}
\value{
The version string returned by \code{GEOSversion();}
}
\description{
GEOS version information
}
\examples{
libgeos_version()

}
