% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libgeos-package.R
\docType{package}
\name{libgeos-package}
\alias{libgeos}
\alias{libgeos-package}
\title{libgeos: Open Source Geometry Engine ('GEOS') C API}
\description{
Provides the Open Source Geometry Engine ('GEOS') as a C API that can be used to write high-performance C and C++ geometry operations using R as an interface. Headers are provided to make linking to and using these functions from C++ code as easy and as safe as possible. This package contains an internal copy of the 'GEOS' library to guarantee the best possible consistency on multiple platforms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paleolimbot.github.io/libgeos/}
  \item \url{https://github.com/paleolimbot/libgeos}
  \item Report bugs at \url{https://github.com/paleolimbot/libgeos/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

Authors:
\itemize{
  \item GEOS authors
}

Other contributors:
\itemize{
  \item Martin Davis [contributor, copyright holder]
  \item Benjamin Campbell [contributor, copyright holder]
  \item Tomasz Sowa [contributor, copyright holder]
  \item Christian Kaiser [contributor, copyright holder]
  \item David Skea [contributor, copyright holder]
  \item Daniel Baston [contributor, copyright holder]
  \item Sandro Santilli [contributor, copyright holder]
  \item Mateusz Loskot [contributor, copyright holder]
  \item Paul Ramsey [contributor, copyright holder]
  \item Olivier Devillers [contributor, copyright holder]
  \item Sean Gillies [contributor, copyright holder]
  \item Mika Heiskanen [contributor, copyright holder]
  \item Safe Software Inc. [copyright holder]
  \item Refractions Research Inc. [copyright holder]
  \item Vivid Solutions Inc. [copyright holder]
}

}
\keyword{internal}
