% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libgeos-package.R
\docType{package}
\name{libgeos-package}
\alias{libgeos}
\alias{libgeos-package}
\title{libgeos: Open Source Geometry Engine (GEOS) C API}
\description{
Provides the Open Source Geometry Engine (GEOS) as a
  C API that can be used to write high-performance C and C++
  geometry operations using R as an interface. Headers are provided
  to make linking to and using these functions from C++ code as
  easy and as safe as possible.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paleolimbot.github.io/libgeos}
  \item \url{https://github.com/paleolimbot/libgeos}
  \item Report bugs at \url{https://github.com/paleolimbot/libgeos/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

Authors:
\itemize{
  \item GEOS authors [copyright holder]
}

Other contributors:
\itemize{
  \item Refractions Research Inc. [copyright holder]
  \item Vivid Solutions Inc. [copyright holder]
}

}
\keyword{internal}
