% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libgeos.R
\docType{package}
\name{libgeos-package}
\alias{libgeos}
\alias{libgeos-package}
\title{libgeos}
\description{
libgeos
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.github.com/rundel/libgeos}
  \item Report bugs at \url{http://www.github.com/rundel/libgeos/issues}
}

}
